requirements.txt
telethon>=1.34.0
google-api-python-client>=2.118.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
